The commands "Export", "Checkout" and "Get Merge Info" of the log dialog did not pass their revision numbers correctly. This is a regression of revision 8011. The SVNRev object needs to be casted to LONG before passing it to the string format function.
git-svn-id: https://svn.code.sf.net/p/tortoisesvn/code/trunk@10440 154be670-7969-47af-9ab0-c6fb378d2295

+++ src/TortoiseProc/LogDlg.cpp
@@ -3549,7 +3549,7 @@ void CLogDlg::ShowContextMenuForRevisions(CWnd* /*pWnd*/, CPoint point)
 				CString url = _T("tsvn:")+pathURL;
 				sCmd.Format(_T("%s /command:export /url:\"%s\" /revision:%ld"),
 					CPathUtils::GetAppDirectory()+_T("TortoiseProc.exe"),
-					url, revSelected);
+					url, (LONG)revSelected);
 				CAppUtils::LaunchApplication(sCmd, NULL, false);
 			}
 			break;
@@ -3559,7 +3559,7 @@ void CLogDlg::ShowContextMenuForRevisions(CWnd* /*pWnd*/, CPoint point)
 				CString url = _T("tsvn:")+pathURL;
 				sCmd.Format(_T("%s /command:checkout /url:\"%s\" /revision:%ld"),
 					CPathUtils::GetAppDirectory()+_T("TortoiseProc.exe"),
-					url, revSelected);
+					url, (LONG)revSelected);
 				CAppUtils::LaunchApplication(sCmd, NULL, false);
 			}
 			break;
@@ -3569,7 +3569,7 @@ void CLogDlg::ShowContextMenuForRevisions(CWnd* /*pWnd*/, CPoint point)
 				CString url = _T("tsvn:")+pathURL;
 				sCmd.Format(_T("%s /command:log /path:\"%s\" /revision:%ld /merge"),
 					CPathUtils::GetAppDirectory()+_T("TortoiseProc.exe"),
-					pathURL, revSelected);
+					pathURL, (LONG)revSelected);
 				CAppUtils::LaunchApplication(sCmd, NULL, false);
 			}
 			break;
